Fix issue with kernel install
When the absolute path to the Python interpreter was not exactly
pointing to an interpreter but a shell script, uv would fail to
find the interpreter. Instead of adding the interpreter to the path
and instruct uv to look for interpreters in the system path, we now
explictly provide the path to the interpreter.

Another issue that happens was if the provided interpreter was a
script that require bash or any shell tools, they could not be
found because the PATH was limited to uv install path and the
interpreter directory. We have added /bin and /usr/bin to PATH.