Introduce parameter for disabling user config
In Jupyterhub < 4.0, disabling the user config excluded disabling
user custom kernelspecs. Since 4.0, the user defined kernels are
also included by the disabling, so we introduce the ability to
allow user defined config to allow custom defined kernels.

This introduces the risk that a badly defined user config mess
with the proper launching of Jupyter, so we keep it to true by
default.